#164 Move Highscores class from `level.save` to `server`
Also renamed it to HighscoreClientAdapter, to give it a more logic name.

<naming,organization
