+++ src/main/java/nl/tudelft/ti2206/group9/gui/scene/HighScoreScene.java
@@ -17,9 +17,9 @@ import javafx.scene.layout.GridPane;
 import nl.tudelft.ti2206.group9.ShaftEscape;
 import nl.tudelft.ti2206.group9.gui.popup.WarningPopup;
 import nl.tudelft.ti2206.group9.level.State;
-import nl.tudelft.ti2206.group9.level.save.Highscores;
-import nl.tudelft.ti2206.group9.level.save.Highscores.Highscore;
-import nl.tudelft.ti2206.group9.level.save.Highscores.ResultCallback;
+import nl.tudelft.ti2206.group9.server.HighscoreClientAdapter;
+import nl.tudelft.ti2206.group9.server.HighscoreClientAdapter.Highscore;
+import nl.tudelft.ti2206.group9.server.HighscoreClientAdapter.ResultCallback;
 import nl.tudelft.ti2206.group9.util.GameObserver;
 
 /**
@@ -140,16 +140,16 @@ public class HighScoreScene extends AbstractMenuScene {
                         GameObserver.Menu.HIGHSCORES_BACK);
                 ShaftEscape.setScene(new MainMenuScene());
             } else if (type == BType.FETCH) {
-                if (Highscores.connect(input.getText())) {
-                    highscoreList = Highscores.getGlobal(SCORE_COUNT, CALLBACK);
+                if (HighscoreClientAdapter.connect(input.getText())) {
+                    highscoreList = HighscoreClientAdapter.getGlobal(SCORE_COUNT, CALLBACK);
                 } else {
                     FAIL_CALLBACK.callback(false);
                 }
             } else if (type == BType.SEND) {
-                if (Highscores.connect(input.getText())) {
-                    Highscores.add(State.getPlayerName(), State.getHighscore(),
+                if (HighscoreClientAdapter.connect(input.getText())) {
+                    HighscoreClientAdapter.add(State.getPlayerName(), State.getHighscore(),
                             FAIL_CALLBACK);
-                    highscoreList = Highscores.getGlobal(SCORE_COUNT, CALLBACK);
+                    highscoreList = HighscoreClientAdapter.getGlobal(SCORE_COUNT, CALLBACK);
                 } else {
                     FAIL_CALLBACK.callback(false);
                 }
